+++ yarn-project/aztec-rpc/src/synchroniser/synchroniser.ts
@@ -1,11 +1,12 @@
+import { AcirSimulator } from '@aztec/acir-simulator';
 import { AztecNode } from '@aztec/aztec-node';
 import { Grumpkin } from '@aztec/barretenberg.js/crypto';
 import { BarretenbergWasm } from '@aztec/barretenberg.js/wasm';
 import { AztecAddress, createDebugLogger, InterruptableSleep } from '@aztec/foundation';
+import { L2BlockContext } from '@aztec/l2-block';
 import { TxHash } from '@aztec/tx';
 import { AccountState } from '../account_state/index.js';
 import { Database, TxDao } from '../database/index.js';
-import { L2BlockContext } from '@aztec/l2-block';
 
 export class Synchroniser {
   private runningPromise?: Promise<void>;
@@ -16,6 +17,7 @@ export class Synchroniser {
   constructor(
     private node: AztecNode,
     private db: Database,
+    private simulator: AcirSimulator,
     private bbWasm: BarretenbergWasm,
     private log = createDebugLogger('aztec:aztec_rpc_synchroniser'),
   ) {}
@@ -73,7 +75,7 @@ export class Synchroniser {
   }
 
   public async addAccount(privKey: Buffer) {
-    this.accountStates.push(new AccountState(privKey, this.db, this.node, new Grumpkin(this.bbWasm)));
+    this.accountStates.push(new AccountState(privKey, this.db, this.simulator, this.node, new Grumpkin(this.bbWasm)));
     await Promise.resolve();
   }
